Life about me | Marinelle Mondejar
Hey guys I'm Marinelle Mondejar. I'm an Filipino entertainer and Fine Arts student from the Philippines.I was born in Quezon City, Philippines. My mom is a housewife. She lives in my city. And my dad is a politician, engineer and businessman and farmer. He lives in rural place. Diagnosed Asperger's syndrome since I was four. I was my first time I heard it when I was 17, because my mom didn't told me when I was a kid. When I was nine, it was my first time experienced as an underdog in school. Many students bullied me, I couldn't count them but I think eight of them were bullies.I like to entertain in front of the people and I also like taking and editing pictures and videos.